Fix config in tests without runner
